requirements.txt
playwright==1.52.0
pytest==7.4.0
pytest-playwright==0.7.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest-xdist==3.7.0
PyYAML==6.0.2
